Fully remove rustc_middle::ich.
