+++ compiler/rustc_middle/src/lib.rs
@@ -77,7 +77,6 @@ pub mod query;
 pub mod arena;
 pub mod dep_graph;
 pub mod hir;
-pub mod ich;
 pub mod infer;
 pub mod lint;
 pub mod middle;
